requirements.txt
Django==1.11.7
djangorestframework==3.7.3
markdown==2.6.9
django-filter==1.1.0
django-debug-toolbar==1.9.1
